Starting SQL bump.

+++ Backend/src/main/java/de/presti/ree6/backend/utils/data/container/CommandStatsContainer.java
@@ -17,6 +17,6 @@ public class CommandStatsContainer {
     }
 
     String command;
-    int uses;
+    long uses;
 
 }
